+++ lang/it-IT.php
@@ -100,3 +100,4 @@ $edit = "Edit";
 $delete = "Elimina";
 $voicemail = "Voicemail";
 $volunteer_has_already_joined_the_call_goodbye = "A volunteer has already joined the call... goodbye";
+$meetings_have_been_texted = "meetings have been texted to you";