+++ lang/pig-latin.php
@@ -92,3 +92,4 @@ $edit = "editcray";
 $delete = "Eleteday";
 $voicemail = "Voicemail";
 $volunteer_has_already_joined_the_call_goodbye = "A volunteer has already joined the call... goodbye";
+$meetings_have_been_texted = "meetings have been texted to you";